requirements.txt
requests
tweepy
pillow